Code modularization
Separé las funciones en varios archivos
